KütüphaneSistemiGüncel/api/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
requests
gunicorn
Flask
PyJWT==2.8.0
